+++ packages/analytics-core/src/types/config/react-native-config.ts
@@ -13,25 +13,27 @@ export type AutocaptureOptionsReactNative = {
   networkTracking?: boolean | NetworkTrackingOptions;
 };
 
-type HiddenOptions = 'apiKey' | 'lastEventId' | 'remoteConfigClient';
+type HiddenOptions = 'apiKey' | 'lastEventId' | 'persistedAppVersion' | 'persistedAppBuild';
 
-export type ReactNativeOptions = Omit<Partial<ReactNativeConfig>, HiddenOptions>;
-
-/* @experimental this config is experimental pending GA of React Native autocapture */
+/** @experimental This config is experimental pending GA of React Native autocapture. */
 export interface ReactNativeAutocaptureOptions {
   sessions?: boolean;
   appLifecycles?: boolean;
+  elementInteractions?: boolean;
+  networkTracking?: boolean | NetworkTrackingOptions;
   screenViews?: boolean;
-  // elementInteractions?
 }
 
 export interface ReactNativeConfig extends Omit<IConfig, 'requestMetadata'> {
   trackingOptions: ReactNativeTrackingOptions;
   /* @deprecated this config is deprecated in favor of config.autocapture */
   trackingSessionEvents?: boolean;
-  autocapture?: boolean | AutocaptureOptionsReactNative;
+  /** @experimental This config is experimental pending GA of React Native autocapture. */
+  autocapture?: boolean | ReactNativeAutocaptureOptions;
   migrateLegacyData?: boolean;
   appVersion?: string;
+  persistedAppVersion?: string;
+  persistedAppBuild?: string;
   attribution?: ReactNativeAttributionOptions;
   deviceId?: string;
   cookieExpiration: number;
@@ -52,11 +54,7 @@ export interface ReactNativeConfig extends Omit<IConfig, 'requestMetadata'> {
   remoteConfigClient?: IRemoteConfigClient;
 }
 
-// TODO: Merge this into ReactNativeConfig once autocapture is GA
-export interface ReactNativeConfigAutocaptureBeta extends ReactNativeConfig {
-  /* @experimental this config is experimental pending GA of React Native autocapture */
-  autocapture?: boolean | ReactNativeAutocaptureOptions;
-}
+export type ReactNativeOptions = Omit<Partial<ReactNativeConfig>, HiddenOptions>;
 
 export interface ReactNativeAttributionOptions {
   disabled?: boolean;