+++ packages/analytics-react-native/src/index.ts
@@ -27,6 +27,8 @@ export const {
 
 export { Revenue, Identify } from '@amplitude/analytics-core';
 
+export { ampCapture } from './amp-capture';
+
 // Export types to maintain backward compatibility with `analytics-types`.
 // In the next major version, only export customer-facing types to reduce the public API surface.
 import * as Types from './types';